+++ pyproject.toml
@@ -103,7 +103,10 @@ dev-docs = [
 
 dev-lint-format = ["ruff==0.2.1",]
 
-dev-test = ["pytest==8.0.1",]
+dev-test = [
+    "pytest==8.0.1",
+    "pytest-qt"
+]
 
 [project.scripts]
 qudi = 'qudi.runnable:main'
