+++ tests/test_modulemanager.py
@@ -20,64 +20,10 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
-from qudi.core import modulemanager,application
-from qudi.core.logger import get_logger
-import numpy as np
-from qudi.util.yaml import yaml_load
+from qudi.core import modulemanager
 import pytest
-from PySide2 import QtCore, QtWidgets
-import weakref
 import coverage
 import os
-
-
-CONFIG = 'C:/qudi/qudi-core/tests/dummy.cfg'
-CONFIG = 'C:/qudi/default.cfg'
-
-
-@pytest.fixture(scope="module")
-def qt_app():
-    app_cls = QtWidgets.QApplication
-    app = app_cls.instance()
-    if app is None:
-        app = app_cls()
-    return app
-
-@pytest.fixture(scope="module")
-def qudi_instance(qt_app):
-    instance = application.Qudi.instance()
-    if instance is None:
-        instance = application.Qudi(config_file=CONFIG)
-    instance_weak = weakref.ref(instance)
-    return instance_weak()
-
-
-@pytest.fixture(scope='module')
-def config():
-    configuration = (yaml_load(CONFIG))
-    return configuration
-
-@pytest.fixture(scope='module')
-def sample_module_gui(config):
-    sample_base = 'gui'
-    sample_module_name, sample_module_cfg = list(config[sample_base].items())[0]
-    return sample_base, sample_module_name, sample_module_cfg
-
-@pytest.fixture(scope='module')
-def sample_module_logic(config):
-    sample_base = 'logic'
-    sample_module_name, sample_module_cfg = list(config[sample_base].items())[0]
-    return sample_base, sample_module_name, sample_module_cfg
-
-@pytest.fixture(scope='module')
-def sample_module_hardware(config):
-    sample_base = 'hardware'
-    sample_module_name, sample_module_cfg = list(config[sample_base].items())[0]
-    return sample_base, sample_module_name, sample_module_cfg
-
-@pytest.fixture(scope="module")
-def module_manager(qudi_instance):
-    return qudi_instance.module_manager
 
 
 @pytest.fixture(autouse=True)
@@ -88,18 +34,16 @@ def coverage_for_each_test(request):
     yield
     # Stop coverage collection
     cov.stop()
-    
     # Create a unique directory for each test function
     test_dir =  f"coverage_{request.node.nodeid.replace('/', '_').replace(':', '_')}"
     os.makedirs(test_dir, exist_ok=True)
     
-    # Save the coverage report
     cov.html_report(directory=test_dir)
     cov.save()
 
     print(f"Coverage report saved to {test_dir}")
 
-'''
+
 def test_add_module( module_manager, sample_module_gui, qtbot):
     """  Test the add_module function if it correctly adds modules
     Parameters
@@ -178,25 +122,37 @@ def test_refresh_module_links(module_manager, sample_module_gui, sample_module_l
     assert gui_module in logic_dependent_modules
 
     gui_module = module_manager._modules.pop(sample_gui_name, None)
-    #gui_module.deactivate()
 
     logic_module = module_manager._modules.pop(sample_logic_name, None)
-    #logic_module.deactivate()
 
-'''
-def test_activate_module(module_manager, config, sample_module_hardware, sample_module_logic, sample_module_gui):
-        for base in ['gui', 'logic', 'hardware']:
-            for module_name, module_cfg in list(config[base].items()):
-                module_manager.add_module(module_name, base, module_cfg, allow_overwrite=False, emit_change=True )
-        
-        _, sample_hardware, _ = sample_module_hardware
-        _, sample_logic, _ = sample_module_logic
-        _, sample_gui, _ = sample_module_gui
-        module_manager.activate_module(sample_gui)
-        print(sample_gui)
-        gui_managed_module = module_manager.modules[sample_gui]
-        hardware_managed_module = module_manager.modules[sample_hardware]
 
-        assert hardware_managed_module.is_active
-        assert gui_managed_module.is_active
+def test_activate_module(module_manager, config, sample_module_hardware, sample_module_logic, sample_module_gui, teardown_modules):
+    """ Tests if modules are actived correctly 
+    Activating gui module should also active corresponding logic and hardware modules
+    Parameters
+    ----------
+    module_manager : fixture
+        fixture for instance of module manager
+    config : fixture
+        fixture for loaded yaml config
+    sample_module_hardware : fixture
+        fixture for instance of managed module for a sample hardware module
+    sample_module_gui : fixture
+        fixture for instance of managed module for a sample gui module
+    sample_module_logic : fixture
+        fixture for instance of managed module for a sample logic module
+    """        
+    for base in ['gui', 'logic', 'hardware']:
+        for module_name, module_cfg in list(config[base].items()):
+            module_manager.add_module(module_name, base, module_cfg, allow_overwrite=False, emit_change=True )
+    
+    _, sample_hardware, _ = sample_module_hardware
+    _, sample_logic, _ = sample_module_logic
+    _, sample_gui, _ = sample_module_gui
+    module_manager.activate_module(sample_gui)
+    gui_managed_module = module_manager.modules[sample_gui]
+    hardware_managed_module = module_manager.modules[sample_hardware]
+
+    assert hardware_managed_module.is_active
+    assert gui_managed_module.is_active
