+++ tests/test_modules_coverage.py
@@ -20,65 +20,18 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
-from qudi.core import modulemanager,application
-from qudi.core.logger import get_logger
-import numpy as np
-from qudi.util.yaml import yaml_load
-import pytest
-from PySide2 import QtWidgets
-import weakref
 import coverage
 import os
 import time
 
-
-CONFIG = 'C:/qudi/default.cfg'
-
-
-@pytest.fixture(scope="module")
-def qt_app():
-    app_cls = QtWidgets.QApplication
-    app = app_cls.instance()
-    if app is None:
-        app = app_cls()
-    return app
-
-@pytest.fixture(scope="module")
-def qudi_instance(qt_app):
-    instance = application.Qudi.instance()
-    if instance is None:
-        instance = application.Qudi(config_file=CONFIG)
-    instance_weak = weakref.ref(instance)
-    return instance_weak()
-
-
-@pytest.fixture(scope='module')
-def config():
-    configuration = (yaml_load(CONFIG))
-    return configuration
-
-
-@pytest.fixture(scope="module")
-def module_manager(qudi_instance):
-    return qudi_instance.module_manager
-
     
-@pytest.fixture(scope='module')
-def sample_module_gui(config):
-    sample_base = 'gui'
-    sample_module_name, sample_module_cfg = list(config[sample_base].items())[0]
-    return sample_base, sample_module_name, sample_module_cfg
-
-
-def test_activate_module(module_manager, config,sample_module_gui):
+def test_coverage_individual(qt_app, module_manager, config):
         for base in ['gui', 'logic', 'hardware']:
             for module_name, module_cfg in list(config[base].items()):
                 module_manager.add_module(module_name, base, module_cfg, allow_overwrite=False, emit_change=True )
-        #_, sample_gui, _ = sample_module_gui
-        #module_manager.activate_module(sample_gui)
         gui_base = 'gui'
         for module_name, _ in list(config[gui_base].items()):
-            print(module_name)
+            #print(module_name)
             cov = coverage.Coverage()
             cov.start()
             
@@ -86,16 +39,18 @@ def test_activate_module(module_manager, config,sample_module_gui):
             module_manager.activate_module(module_name)
             cov.stop()
             assert module_manager.modules[module_name].is_active
-            
             # Create a unique directory for each test function
             test_dir =  os.path.join('coverage',f"coverage_{module_name} qudi iqo")
             os.makedirs(test_dir, exist_ok=True)
             
             # Save the coverage report
             cov.html_report(directory=test_dir)
+            cov.annotate(directory=test_dir)
+
             cov.save()
 
-            print(f"Coverage report saved to {test_dir}")
-
+            #print(f"Coverage report saved to {test_dir}")
             #module_manager.deactivate_module(module_name)
+
+
